Version 0.0.2
Networksystem; Chunk and Block reciving

+++ src/com/minecraft/world/Block/BlockManager.java
@@ -15,12 +15,12 @@ import java.util.logging.Logger;
  * @author Pascal
  */
 public class BlockManager {
-    private HashMap<Integer,Class<?>> registeredBlocks = new HashMap<Integer,Class<?>>();
-    public void register(int id,Class<?> block){
+    private HashMap<Short,Class<?>> registeredBlocks = new HashMap<Short,Class<?>>();
+    public void register(short id,Class<?> block){
         if(registeredBlocks.containsKey(id))throw new IllegalArgumentException("Block is already registered");
         registeredBlocks.put(id, block);
     }
-    public Block createBlock(int id){
+    public Block createBlock(short id){
         try {
             return (Block)registeredBlocks.get(id).newInstance();
         } catch (Exception ex) {
